requirements.txt
django
pymysql
djangorestframework
django-cors-headers
python-decouple
cryptography
colorlog